fix attribute error

+++ src/main.py
@@ -95,11 +95,9 @@ def import_dicom_volumes(
                 w.workflow_output(api, project.id, "project")
 
         if g.REMOVE_SOURCE and not g.IS_ON_AGENT:
-            if g.INPUT_DIR is not None:
-                path_to_remove = g.INPUT_DIR
-            else:
-                path_to_remove = g.INPUT_FILE
-            api.file.remove(team_id=g.TEAM_ID, path=path_to_remove)
+            path_to_remove = g.INPUT_DIR or g.INPUT_FILE
+            if path_to_remove:
+                api.file.remove(team_id=g.TEAM_ID, path=path_to_remove)
             source_dir_name = path_to_remove.lstrip("/").rstrip("/")
             sly.logger.info(msg=f"Source directory: '{source_dir_name}' was successfully removed.")
